# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RN CF API",
    "version": "0.0.1"
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "unauthorized"
          },
          "message": {
            "type": "string",
            "example": "Authentication required"
          }
        },
        "required": ["error", "message"]
      },
...
      "AccountDeletedResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [true]
          }
        },
        "required": ["deleted"]
      },
      "DeleteAccountRequest": {
        "type": "object",
        "properties": {
          "confirmation": {
            "type": "string",
            "enum": ["DELETE"]
          }
        },
        "required": ["confirmation"]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"],
            "example": "ok"
          },
          "service": {
            "type": "string",
            "example": "rn-cf-api"
          }
        },
        "required": ["status", "service"]
      },
...
      "OrganizationContextResponse": {
        "type": "object",
        "properties": {
          "organization": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "org_123"
              },
              "role": {
                "type": "string",
                "enum": ["owner", "admin", "member"]
              }
            },
            "required": ["id", "role"]
          }
        },
        "required": ["organization"]
      }
    },
...
  },
  "paths": {
...
    "/api/account": {
      "delete": {
        "operationId": "deleteAccount",
        "tags": ["Account"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Local account data deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Verified account required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Active store or Stripe subscription must be managed first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/private/pro": {
      "get": {
        "operationId": "getProPing",
        "tags": ["Billing"],
        "responses": {
          "200": {
            "description": "Capability protected by the Pro entitlement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Pro entitlement required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/private/organization": {
      "get": {
        "operationId": "getOrganizationContext",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-organization-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Verified organization context for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationContextResponse"
                }
              }
            }
          },
          "400": {
            "description": "No organization was selected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The user is not a member of the selected organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}