# packages/api-client/src/generated/models/deleteAccountRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * RN CF API
 * OpenAPI spec version: 0.0.1
 */
import type { DeleteAccountRequestConfirmation } from "./deleteAccountRequestConfirmation";

export interface DeleteAccountRequest {
  confirmation: DeleteAccountRequestConfirmation;
}